+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog #
 
 ## Unreleased ##
+ * **BREAKING** Authenticator `label` has to be unique for user. This can potentialy break if you have multiple tokens for user.
 
 ## 0.15 ##
  * Add ``DJANGO_FIDO_RP_NAME`` setting.